requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_tflow_select=2.3.0=mkl
absl-py=0.10.0=py37hc8dfbb8_1
argon2-cffi=20.1.0=py37h8f50634_2
astor=0.8.1=pyh9f0ad1d_0
async_generator=1.10=py_0
attrs=20.2.0=pyh9f0ad1d_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.1=py_0
blas=1.0=mkl
bleach=3.2.1=pyh9f0ad1d_0
brotlipy=0.7.0=py37hb5d75c8_1001
c-ares=1.16.1=h516909a_3
ca-certificates=2020.10.14=0
cachetools=4.1.1=pypi_0
certifi=2020.6.20=py37_0
cffi=1.14.3=py37he30daa8_0
chardet=3.0.4=py37he5f6b98_1008
click=7.1.2=pyh9f0ad1d_0
conllu=4.2.1=pypi_0
cryptography=3.1.1=py37hff6837a_1
cudatoolkit=10.1.243=h6bb024c_0
cycler=0.10.0=py_2
datasets=1.1.2=pypi_0
dbus=1.13.18=hb2f20db_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
dill=0.3.3=pypi_0
elasticsearch=7.9.1=pypi_0
emoji=0.6.0=pypi_0
entrypoints=0.3=py37hc8dfbb8_1002
expat=2.2.10=he6710b0_2
filelock=3.0.12=pyh9f0ad1d_0
fire=0.3.1=pypi_0
fontconfig=2.13.0=h9420a91_0
freetype=2.10.3=h5ab3b9f_0
fsspec=0.8.4=pypi_0
future=0.18.2=pypi_0
gast=0.2.2=py_0
gitdb=4.0.5=pypi_0
gitpython=3.1.11=pypi_0
glib=2.56.2=hd408876_0
google-auth=1.23.0=pypi_0
google-auth-oauthlib=0.4.2=pypi_0
google-pasta=0.2.0=pyh8c360ce_0
gperftools=2.7=h767d802_2
grpcio=1.31.0=py37hb0870dc_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
h5py=2.10.0=nompi_py37hf7afa78_105
hdf5=1.10.6=nompi_h3c11f04_101
icu=58.2=he6710b0_3
idna=2.10=pyh9f0ad1d_0
importlib-metadata=2.0.0=py_1
importlib_metadata=2.0.0=1
iniconfig=1.1.1=pypi_0
intel-openmp=2020.2=254
intelpython=2020.2=0
ipykernel=5.3.4=py37h5ca1d4c_0
ipython=7.18.1=py37hc6149b9_1
ipython_genutils=0.2.0=py_1
jedi=0.17.2=py37hc8dfbb8_1
jinja2=2.11.2=pyh9f0ad1d_0
joblib=0.17.0=py_0
jpeg=9b=h024ee3a_2
json5=0.9.5=pyh9f0ad1d_0
jsonschema=3.2.0=py_2
jupyter_client=6.1.7=py_0
jupyter_core=4.6.3=py37hc8dfbb8_2
jupyterlab=2.2.8=py_0
jupyterlab_pygments=0.1.2=pyh9f0ad1d_0
jupyterlab_server=1.2.0=py_0
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.0=py_0
kiwisolver=1.2.0=py37h99015e2_1
krb5=1.17.1=hfafb76e_3
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libcurl=7.71.1=hcdd3856_8
libedit=3.1.20191231=h14c3975_1
libev=4.33=h516909a_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.5.0=hae1eefd_17
libgfortran4=7.5.0=hae1eefd_17
libnghttp2=1.41.0=h8cfc5f6_2
libpng=1.6.37=hbc83047_0
libprotobuf=3.13.0.1=h8b12597_0
libsodium=1.0.18=h516909a_1
libssh2=1.9.0=hab1572f_5
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h1bed415_2
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
lz4-c=1.9.2=heb0550a_3
markdown=3.3.1=pyh9f0ad1d_0
markupsafe=1.1.1=py37hb5d75c8_2
matplotlib=3.1.2=py37h4fdacc2_0
matplotlib-base=3.3.2=py37hc9afd2a_1
mistune=0.8.4=py37h8f50634_1002
mkl=2020.2=256
mkl-service=2.3.0=py37he904b0f_0
mkl_fft=1.2.0=py37h23d657b_0
mkl_random=1.1.1=py37h0573a6f_0
multiprocess=0.70.11.1=pypi_0
nbclient=0.5.1=py_0
nbconvert=6.0.7=py37hc8dfbb8_1
nbformat=5.0.8=py_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.4.1=py_0
ninja=1.10.1=py37hfd86e86_0
nltk=3.5=pypi_0
notebook=6.1.4=py37hc8dfbb8_1
numpy=1.19.1=py37hbc911f0_0
numpy-base=1.19.1=py37hfa32c7d_0
oauthlib=3.1.0=pypi_0
olefile=0.46=py37_0
openssl=1.1.1h=h7b6447c_0
opt_einsum=3.3.0=py_0
packaging=20.4=pyh9f0ad1d_0
pandas=1.1.3=py37he6710b0_0
pandoc=2.11.0.2=hd18ef5c_0
pandocfilters=1.4.2=py_1
parso=0.7.1=pyh9f0ad1d_0
pcre=8.44=he6710b0_0
perl=5.30.3=h516909a_1
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=7.2.0=py37hb39fc2d_0
pip=20.2.3=py37_0
pluggy=0.13.1=pypi_0
portalocker=2.0.0=pypi_0
prometheus_client=0.8.0=pyh9f0ad1d_0
prompt-toolkit=3.0.8=py_0
protobuf=3.13.0.1=py37he6710b0_1
ptyprocess=0.6.0=py_1001
py=1.9.0=pypi_0
pyarrow=2.0.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.20=pyh9f0ad1d_2
pygments=2.7.1=py_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=pyh9f0ad1d_0
pyqt=5.9.2=py37h22d08a2_1
pyrsistent=0.17.3=py37h8f50634_1
pysocks=1.7.1=py37he5f6b98_2
pytest=6.1.2=pypi_0
python=3.7.9=h7579374_0
python-dateutil=2.8.1=py_0
python_abi=3.7=1_cp37m
pytorch=1.6.0=py3.7_cuda10.1.243_cudnn7.6.3_0
pytorch-lightning=0.9.0=pypi_0
pytz=2020.1=py_0
pyyaml=5.3.1=pypi_0
pyzmq=19.0.2=py37hac76be4_2
qt=5.9.7=h5867ecd_1
readline=8.0=h7b6447c_0
regex=2020.10.11=py37h8f50634_0
requests=2.24.0=pyh9f0ad1d_0
requests-oauthlib=1.3.0=pypi_0
rouge-score=0.0.4=pypi_0
rsa=4.6=pypi_0
sacrebleu=1.4.14=pypi_0
sacremoses=0.0.43=pyh9f0ad1d_0
scikit-learn=0.23.2=py37h0573a6f_0
scipy=1.5.2=py37h0b6359f_0
seaborn=0.11.0=py_0
send2trash=1.5.0=py_0
sentencepiece=0.1.92=py37h99015e2_0
setuptools=50.3.0=py37hb0f4dca_1
sip=4.19.24=py37he6710b0_0
six=1.15.0=py_0
smmap=3.0.4=pypi_0
sqlite=3.33.0=h62c20be_0
tensorboard=2.2.0=pypi_0
tensorboard-plugin-wit=1.7.0=pypi_0
tensorflow=1.15.0=mkl_py37h28c19af_0
tensorflow-base=1.15.0=mkl_py37he1670d9_0
tensorflow-estimator=1.15.1=pyh2649769_0
termcolor=1.1.0=py_2
terminado=0.9.1=py37hc8dfbb8_1
testpath=0.4.4=py_0
threadpoolctl=2.1.0=py37h6447541_2
tk=8.6.10=hbc83047_0
tokenizers=0.8.1=py37h4c3b196_0
toml=0.10.2=pypi_0
torchvision=0.7.0=py37_cu101
tornado=6.0.4=py37h8f50634_2
tqdm=4.49.0=pypi_0
traitlets=5.0.5=py_0
transformers=3.3.1=py_0
urllib3=1.25.10=py_0
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
werkzeug=0.16.1=py_0
wheel=0.35.1=py_0
wordcloud=1.8.0=py37h8f50634_0
wordsegment=1.3.1=pypi_0
wrapt=1.12.1=py37h8f50634_1
xxhash=2.0.0=pypi_0
xz=5.2.5=h7b6447c_0
zeromq=4.3.3=he1b5a44_2
zipp=3.3.0=py_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
